[Bugfix] CAT2000 contained unnecessary files
The CAT2000 dataset comes with additional saliency maps (I think AIM?)
in subdirectories. For CAT2000 test they already have been removed
as part of the pysaliency import, but for the train dataset this
has been forgotten. This is now fixed, also there is a test for
CAT2000_train v1.1 added.

Signed-off-by: Matthias Kümmmerer <matthias.kuemmerer@bethgelab.org>