Enhancement: HDF5 models better handle loading for subset of keys
so far, loading a model of a subset of keys in the HDF5 file only
worked when the prefix in the HDF5 keys was also part of
the filenames in the stimuli. This often is not the case when
saving multiple datasets in one HDF5 file and should now be handled
better.

Also we now have better exceptions and more tests for HDF5 models.

Signed-off-by: Matthias Kümmerer <matthias@matthias-k.org>