Feature: linear constraints in pysaliency.optpy
So far, the optpy wrapper around scipy.minimize only supported
nonlinear constraints specified as dictionary. In the case
of linear constraints, this can result in bad optimization
performance, because it takes time for the hessian update
strategy to converge to zero.

This comment adds a new LinearConstraint class that allows
to specify a linear constraint via the matrices operating on
the different parameters and which is converted into a
scipy.optimize.LinearConstraint.

Signed-off-by: Matthias Kümmerer <matthias@matthias-k.org>