Use LinearConstraints in saliency_map_conversation_torch
In scipy 1.15.0 a bug was fixed
(https://github.com/scipy/scipy/issues/21193): before that
multiple constraints accidentally shared the hessian update
strategy. This is usually bad, but in our case apparently
it helped optimization because nearly all constraints
are actually linear. Since scipy 1.15.0 the optimization
ran much slower, resulting in failed tests. I reduced the
problem to said scipy bug (thanks to git bisect) and
concluded that I should try to properly implement the
constraints as linear constraints where possible, to make
sure that we have correct hessian information. This is
being done is the present commit and it not only makes
optimization fast again, it also slighly improves
performance in some cases.

Eventually, this should also be done in the respective
theano implementation, but even better might be to finally
get rid of it.

Signed-off-by: Matthias Kümmerer <matthias@matthias-k.org>